simplifying 'stopping docker' in the instructions
FWIW - the docker machine setup includes a one-stop `Docker Quickstart Terminal` that makes a "default" docker-machine and opens a bash shell for you to use. Most of the top segment of these instructions could be replaced by using that directly, but I didn't want to wipe out all the detail there.